Enhancement: Turn method on Format into named constructor on Indent
