+++ test/Unit/Format/FormatTest.php
@@ -227,111 +227,6 @@ final class FormatTest extends Framework\TestCase
     }
 
     /**
-     * @dataProvider providerPureIndentAndSniffedIndent
-     * @dataProvider providerMixedIndentAndSniffedIndent
-     *
-     * @param string $indent
-     * @param string $sniffedIndent
-     */
-    public function testFromJsonReturnsFormatWithIndentSniffedFromArray(string $indent, string $sniffedIndent): void
-    {
-        $json = Json::fromEncoded(
-            <<<JSON
-[
-"foo",
-${indent}"bar",
-    {
-        "qux": "quux"
-    }
-]
-JSON
-        );
-
-        $format = Format::fromJson($json);
-
-        $this->assertInstanceOf(Format::class, $format);
-        $this->assertSame($sniffedIndent, $format->indent()->__toString());
-    }
-
-    /**
-     * @dataProvider providerPureIndentAndSniffedIndent
-     * @dataProvider providerMixedIndentAndSniffedIndent
-     *
-     * @param string $indent
-     * @param string $sniffedIndent
-     */
-    public function testFromJsonReturnsFormatWithIndentSniffedFromObject(string $indent, string $sniffedIndent): void
-    {
-        $json = Json::fromEncoded(
-<<<JSON
-{
-"foo": 9000,
-${indent}"bar": 123,
-    "baz": {
-        "qux": "quux"
-    }
-}
-JSON
-        );
-
-        $format = Format::fromJson($json);
-
-        $this->assertInstanceOf(Format::class, $format);
-        $this->assertSame($sniffedIndent, $format->indent()->__toString());
-    }
-
-    public function providerPureIndentAndSniffedIndent(): \Generator
-    {
-        $characters = [
-            'space' => ' ',
-            'tab' => "\t",
-        ];
-
-        $sizes = [1, 3];
-
-        foreach ($characters as $style => $character) {
-            foreach ($sizes as $size) {
-                $key = \sprintf(
-                    '%s-%d',
-                    $style,
-                    $size
-                );
-
-                $pureIndent = \str_repeat(
-                    $character,
-                    $size
-                );
-
-                yield $key => [
-                    $pureIndent,
-                    $pureIndent,
-                ];
-            }
-        }
-    }
-
-    public function providerMixedIndentAndSniffedIndent(): \Generator
-    {
-        $mixedIndents = [
-            'space-and-tab' => [
-                " \t",
-                ' ',
-            ],
-            'tab-and-space' => [
-                "\t ",
-                "\t",
-            ],
-        ];
-
-        foreach ($mixedIndents as $key => [$mixedIndent, $sniffedIndent]) {
-            yield $key => [
-                $mixedIndent,
-                $sniffedIndent,
-            ];
-        }
-    }
-
-    /**
      * @dataProvider providerEncodedWithoutIndent
      *
      * @param string $encoded
